move widgets in a separate folder
